requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
cloudpickle==1.6.0
colorama==0.4.4
Django==3.1.7
django-rest-framework==0.1.0
djangorestframework==3.12.2
importlib-metadata==3.7.3
iniconfig==1.1.1
joblib==1.0.1
llvmlite==0.36.0
numba==0.53.0
numpy==1.19.5
packaging==20.9
pandas==1.1.5
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.24.1
scipy==1.5.4
shap==0.39.0
six==1.15.0
slicer==0.0.7
sqlparse==0.4.1
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.59.0
typing-extensions==3.7.4.3
zipp==3.4.1
